requirements.txt
langchain==0.3.27
langchain-google-genai
chromadb
python-dotenv
google-generativeai
